Update index.html
"Take into consideration that there might have been some confusion with the WebViewer component and the WebBridge component in Thunkable when designing the script. We are not using the WebViewer component at all. We only have the JS Web Bridge component in the app. So I want you to give the script an update that caters to the Web Bridge component in Thunkable so that it might actually work for Android devices since it works for iOS only so far." -Me

"We need to replace window.ReactNativeWebView.postMessage(msg) with window.postMessage(msg, "*"), which is what Thunkable expects for both Android and iOS Web Bridge communication.

Here is your updated, Android-compatible version:

Updated Script (Web Bridge Only, Cross-Platform)" -ChatGPT